requirements.txt
# lib dependencies
attrs~=25.3.0
cattrs~=24.1.3
pynacl~=1.5.0

# for auth-callout sample
nats-py~=2.10.0